swagger: Fix type of array of Y values in XY SeriesModel
The values are actually of type doubles and not long.

Fixes:
https://github.com/eclipse-cdt-cloud/trace-server-protocol/issues/101

+++ trace-server/org.eclipse.tracecompass.incubator.trace.server.jersey.rest.core/src/org/eclipse/tracecompass/incubator/internal/trace/server/jersey/rest/core/model/SeriesModel.java
@@ -49,7 +49,7 @@ public interface SeriesModel {
      */
     @JsonProperty("yValues")
     @ArraySchema(arraySchema = @Schema(description = "Series' Y values", required = true))
-    long[] getYValues();
+    double[] getYValues();
 
     /**
      * @return The series style.
